+++ packages/runtime-web/src/components/src/utils/sleep.ts
@@ -1,0 +1,5 @@
+export const sleep = (time) => {
+  return new Promise((resolve) => {
+    setTimeout(resolve, time)
+  })
+}